Create an ubertooth-scanner wrapper script
The ubertooth one device is only capable of BTLE scanning in Kismet right now, and we need to find nearby BT devices that are not discoverable. This script wraps around the ubertooth-scanner program (which can find undiscoverable BT devices), and formats that tool's output into JSON so we can pull it into other programs.

Eventually I'd like to have something like this running as a daemon and then invoking Kismet's API to add the devices we find with this to Kismet, but this will work for now.