Create a middleware file for data processing.
Currently we scrape the kismet API every so many seconds for new devices, which isn't that efficient. This middleware is a persistent separate process that can receive messages via UNIX socket and process it. Used in tandem with a kismet plugin forwarding NEW_DEVICE events to the created UNIX socket, this enables data processing that's event driven instead of time driven.