Rework script output format and add HW serial num
Change how the script outputs the formatted data to include the ubertooth's microcontroller serial #, which can work like a UUID in cases where multiple ubertooth devices are deployed. Also fixup the output so it is fully JSON compatible, and add in a call to python's mjson.tool to print a human readable form to terminal.