Add type field to kismet messages and fix ordering
A required field that identified what the type of device kismet was reporting was missing, this adds it. Also changes the order of the fields so that the reported order is device type, device name, device macaddr, device seenby data, instead of the previous order which was the reverse of this. The ordering of the fields in the subscribe message is the opposite of how they will be ordered in eventbus messages.