Config files are at /etc/bluemon, update program to reflect this.
Config files, including the main config file and the zones config file should be housed under /etc/bluemon so they're all in the same place. This updates the program to read configs from there, and also updates the install script to create the directory and populate it with sample configs during the install.