Update install script to match current deployment needs.
There was an old pip install statement in the install script that no longer applies. Comment it out and mark the section for later when we figure out how to do venv with the program. Also fix an issue where kismet might not be able to access serial devices due to not being in the dialout group on Ubuntu.