Make string matches case-insensitive and resolve a logic error.
String comparisons were previously case-sensitive, which could affect device matching logic and cause false negatives for device matches or zone lookups. This resolves that by using python functions to convert the case on both sides of the comparison so they are the same. This also resolves a logic error in device comparison code which was previously missed and fixes an incorrect comment.